Align specs with current analyzers
Keep actions in behavioral specifications, adopt should-prefixed facts, and use concise empty record declarations so the current analyzer set builds cleanly.
