+++ Source/DotNET/Fundamentals.Specs/Json/for_ConceptAsJsonConverter/given/converter_for_converting_to_json.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System.Text;
 using System.Text.Json;
 
 namespace Cratis.Json.for_ConceptAsJsonConverter.given;
@@ -11,8 +10,8 @@ public abstract class converter_for_converting_to_json<TConcept, TUnderlying> : 
     protected ConceptAsJsonConverter<TConcept> converter;
     protected TConcept input;
     protected string result;
-    MemoryStream stream;
-    Utf8JsonWriter writer;
+    protected MemoryStream stream;
+    protected Utf8JsonWriter writer;
 
     protected abstract TUnderlying Expected { get; }
     protected abstract string FormattedExpected { get; }
@@ -26,13 +25,6 @@ public abstract class converter_for_converting_to_json<TConcept, TUnderlying> : 
                     .Invoke([Expected]);
         stream = new();
         writer = new(stream);
-    }
-
-    void Because()
-    {
-        converter.Write(writer, input, default);
-        writer.Flush();
-        result = Encoding.UTF8.GetString(stream.ToArray());
     }
 
     protected void ShouldConvertToJson() => result.ShouldEqual(FormattedExpected);
