+++ Source/DotNET/Fundamentals.Specs/Json/for_ConceptAsJsonConverter/when_converting_date_only_concept_from_json.cs
@@ -1,6 +1,9 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System.Text;
+using System.Text.Json;
+
 namespace Cratis.Json.for_ConceptAsJsonConverter;
 
 public class when_converting_date_only_concept_from_json : given.converter_for_converting_from_json<DateOnlyConcept, DateOnly>
@@ -8,6 +11,17 @@ public class when_converting_date_only_concept_from_json : given.converter_for_c
     protected override DateOnly InputValue => DateOnly.FromDateTime(DateTime.UtcNow);
 
     protected override string FormattedInput => $"\"{input:O}\"";
+
+    void Because()
+    {
+        var fullJson = $"{{ \"prop\": {FormattedInput} }}";
+
+        Utf8JsonReader reader = new(Encoding.UTF8.GetBytes(fullJson).AsSpan());
+        reader.Read();  // Start object
+        reader.Read();  // Property
+        reader.Read();  // Value
+        result = converter.Read(ref reader, ConceptType, default);
+    }
 
     [Fact] void should_convert_to_correct_date_only() => ShouldConvertToCorrectConcept();
 }
