+++ Source/DotNET/Fundamentals.Specs/Json/for_ConceptAsJsonConverter/when_converting_date_only_concept_to_json.cs
@@ -1,12 +1,21 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System.Text;
+
 namespace Cratis.Json.for_ConceptAsJsonConverter;
 
 public class when_converting_date_only_concept_to_json : given.converter_for_converting_to_json<DateOnlyConcept, DateOnly>
 {
     protected override DateOnly Expected => DateOnly.FromDateTime(DateTime.UtcNow);
     protected override string FormattedExpected => $"\"{input.Value:O}\"";
+
+    void Because()
+    {
+        converter.Write(writer, input, default);
+        writer.Flush();
+        result = Encoding.UTF8.GetString(stream.ToArray());
+    }
 
     [Fact] void should_convert_to_correct_date_only() => ShouldConvertToJson();
 }
