+++ Source/DotNET/Fundamentals.Specs/Json/for_ConceptAsJsonConverter/when_converting_decimal_concept_to_json.cs
@@ -1,6 +1,8 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System.Text;
+
 namespace Cratis.Json.for_ConceptAsJsonConverter;
 
 public class when_converting_decimal_concept_to_json : given.converter_for_converting_to_json<DecimalConcept, decimal>
@@ -8,6 +10,13 @@ public class when_converting_decimal_concept_to_json : given.converter_for_conve
     protected override decimal Expected => 42.43M;
 
     protected override string FormattedExpected => "42.43";
+
+    void Because()
+    {
+        converter.Write(writer, input, default);
+        writer.Flush();
+        result = Encoding.UTF8.GetString(stream.ToArray());
+    }
 
     [Fact] void should_convert_to_correct_decimal() => ShouldConvertToJson();
 }
